backend/requirements.txt
Django==5.1.7
Pillow==11.1.0
djangorestframework==3.16.0
djoser==2.3.1
psycopg==3.2.7
firebase-admin==6.2.0